requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
numpy
pytz
firebase-admin 
python-jose
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
APScheduler
sendgrid